requirements.txt
streamlit
openai
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
python-dotenv
pypdf
tiktoken
faiss-cpu